+++ test/ref_sheeted_fixture_test.exs
@@ -1,5 +1,5 @@
 defmodule SvgSpriteEx.RefSheetedFixture do
-  use SvgSpriteEx
+  use SvgSpriteEx.Ref
 
   def ui_ref, do: sprite_ref("regular/xmark", sheet: " UI Actions ")
   def duplicate_ref, do: sprite_ref("regular/xmark", sheet: "ui_actions")